pyproject.toml
[project]
name = "pad-salience-annotations"
version = "0.1.0"
description = "PAD Salience Annotation System with Experiment Management"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.19.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.128.0",
    "jinja2>=3.1.0",
    "pad-analytics",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0",
    "uvicorn>=0.40.0",
]

[tool.uv.sources]
pad-analytics = { git = "https://github.com/PaperAnalyticalDeviceND/pad-analytics.git", rev = "aee1fd12c5845fb0b523592d79b8ea97d988d2aa" }
